Make c10d_functional ops call into _c10d_functional ops
This PR removes the legacy impls of c10d_functional ops which are now irrelevant. For backward compatibility purpose, c10d_functional ops now call into _c10d_functional ops.

We also changed c10d_functional ops to be CompositeExplicitAutograd, so that when traced, only _c10d_functional ops appear in the graph. After this, we'll be able to remove the Inductor IR for the legacy functional collectives.

ghstack-source-id: 4f6b692d1a598f66da1f69fa06585782188b9908
Pull Request resolved: https://github.com/pytorch/pytorch/pull/124979